+++ client/dev.sh
@@ -11,4 +11,9 @@ fi
 ln -sf $(pwd)/dist ../public/bin/
 ln -sf $(pwd)/../front/src/app.js ../public/bin/proteinpaint.js
 
+# needed to track messages for browser notification
+if [[ ! -d ../.sse/messages ]]; then
+	mkdir -p ../.sse/messages
+fi
+
 ENV=dev node esbuild.config.mjs