+++ client/plots/sampleScatter.js
@@ -41,8 +41,8 @@ NOTE
 "sample" and "category" attributes here are hardcoded
 
 */
-export const minDotSize = 9
-export const maxDotSize = 300
+export const minShapeSize = 9
+export const maxShapeSize = 300
 class Scatter {
 	constructor() {
 		this.type = 'sampleScatter'
@@ -264,31 +264,31 @@ class Scatter {
 			label: 'Shape',
 			vocabApi: this.app.vocabApi
 		}
-		const dotSizeOption = {
+		const shapeSizeOption = {
 			label: 'Sample area',
 			type: 'number',
 			chartType: 'sampleScatter',
 			settingsKey: 'size',
-			title: 'Sample area, in square pixels',
+			title: 'Sample size, squared gives you the sample area',
 			min: 0
 		}
-		const minDotSizeOption = {
-			label: 'Min area',
+		const minShapeSizeOption = {
+			label: 'Min size',
 			type: 'number',
 			chartType: 'sampleScatter',
-			settingsKey: 'minDotSize',
-			title: 'Minimum area, in square pixels',
-			min: minDotSize,
-			max: maxDotSize
+			settingsKey: 'minShapeSize',
+			title: 'Minimum sample size',
+			min: minShapeSize,
+			max: maxShapeSize
 		}
-		const maxDotSizeOption = {
-			label: 'Max area',
+		const maxShapeSizeOption = {
+			label: 'Max size',
 			type: 'number',
 			chartType: 'sampleScatter',
-			settingsKey: 'maxDotSize',
-			title: 'Maximum area, in square pixels',
-			min: minDotSize,
-			max: maxDotSize
+			settingsKey: 'maxShapeSize',
+			title: 'Maximum sample size',
+			min: minShapeSize,
+			max: maxShapeSize
 		}
 		const orientation = {
 			label: 'Scale order',
@@ -443,12 +443,12 @@ class Scatter {
 				inputs.splice(4, 0, shapeOption)
 				inputs.splice(5, 0, scaleDotOption)
 				if (this.config.scaleDotTW) {
-					inputs.splice(6, 0, minDotSizeOption)
-					inputs.splice(7, 0, maxDotSizeOption)
+					inputs.splice(6, 0, minShapeSizeOption)
+					inputs.splice(7, 0, maxShapeSizeOption)
 					inputs.splice(8, 0, orientation)
 					if (hasRef) inputs.splice(9, 0, refSizeOption)
 				} else {
-					inputs.splice(6, 0, dotSizeOption)
+					inputs.splice(6, 0, shapeSizeOption)
 					if (hasRef) inputs.splice(7, 0, refSizeOption)
 				}
 
@@ -491,12 +491,12 @@ class Scatter {
 			inputs.splice(2, 0, shapeOption)
 			inputs.splice(3, 0, scaleDotOption)
 			if (this.config.scaleDotTW) {
-				inputs.splice(4, 0, minDotSizeOption)
-				inputs.splice(5, 0, maxDotSizeOption)
+				inputs.splice(4, 0, minShapeSizeOption)
+				inputs.splice(5, 0, maxShapeSizeOption)
 				inputs.splice(6, 0, orientation)
 				if (hasRef) inputs.splice(7, 0, refSizeOption)
 			} else {
-				inputs.splice(4, 0, dotSizeOption)
+				inputs.splice(4, 0, shapeSizeOption)
 				if (hasRef) inputs.splice(5, 0, refSizeOption)
 			}
 			inputs.push(showAxes)
@@ -651,11 +651,11 @@ export function makeChartBtnMenu(holder, chartsInstance) {
 
 export function getDefaultScatterSettings() {
 	return {
-		size: 16,
-		minDotSize: 9,
-		maxDotSize: 144,
+		size: 4,
+		minShapeSize: 3,
+		maxShapeSize: 12,
 		scaleDotOrder: 'Ascending',
-		refSize: 9,
+		refSize: 3,
 		svgw: 500,
 		svgh: 500,
 		svgd: 500,